+++ src/lib/components/lemmy/user/UserCard.svelte
@@ -19,12 +19,11 @@
     import Markdown from '$lib/components/markdown/Markdown.svelte'
     import Menu from '$lib/components/ui/menu/Menu.svelte'
     import MenuButton from '$lib/components/ui/menu/MenuButton.svelte'
-    import Modal from '$lib/components/ui/modal/Modal.svelte'
     import RelativeDate from '$lib/components/util/RelativeDate.svelte'
     import SidebarFooter from '$lib/components/ui/SidebarFooter.svelte';
     import StickyCard from '$lib/components/ui/StickyCard.svelte'
-    import TextArea from '$lib/components/input/TextArea.svelte'
     import UserLink from '$lib/components/lemmy/user/UserLink.svelte'
+    import UserSendMessageModal from './UserSendMessageModal.svelte'
 
     import {
         Cake,
@@ -34,16 +33,15 @@
         Hashtag,
         Home,
         Icon,
-        InformationCircle,
         NoSymbol,
         Newspaper,
         PencilSquare,
         Share,
         ShieldCheck,
         ShieldExclamation,
-        Trophy,
         UserCircle,
     } from 'svelte-hero-icons'
+    
     
     
     export let person: PersonView | LocalUserView
@@ -55,9 +53,7 @@
     if ((person as PersonView).is_admin) is_admin=true
     
     let blocking = false
-    let loadingMessage = false
     let messaging = false
-    let message = ''
 
     async function blockUser(block: number) {
         if (!$profile?.user || !$profile?.jwt) throw new Error('Unauthenticated')
@@ -81,6 +77,7 @@
             toast({
                 content: `Successfully ${blocked ? 'unblocked' : 'blocked'} that user.`,
                 type: 'success',
+                title: 'Blocked'
             })
 
             goto($page.url, {
@@ -90,67 +87,20 @@
             toast({
                 content: err as any,
                 type: 'error',
+                title: 'Error'
             })
         }
         blocking = false
     }
 
-    async function sendMessage() {
-        if (!$profile?.jwt || message == '') return
-        
-        loadingMessage = true
-
-        try {
-            await getClient().createPrivateMessage({
-                content: message,
-                recipient_id: person.person.id,
-            })
-
-            toast({
-                content: 'Successfully sent that person a message.',
-                type: 'success',
-            })
-
-            messaging = false
-        } catch (err) {
-            toast({
-                content: err as any,
-                type: 'error',
-            })
-        }
-
-        loadingMessage = false
-    }
 </script>
 
 {#if display}
     
     <!---DM Compose Modal--->
     {#if $profile?.user}
-        <Modal bind:open={messaging} title="Message">
-            <form on:submit|preventDefault={sendMessage} class="flex flex-col gap-4">
-                <p class="inline-flex flex-row gap-2 items-center">
-                    Sending <UserLink avatar user={person.person} /> a message
-                </p>
-                <TextArea
-                    bind:value={message}
-                    label="Message"
-                    rows={8}
-                />
-                <Button
-                    color="primary"
-                    size="lg"
-                    submit
-                    loading={loadingMessage}
-                    disabled={loadingMessage}
-                >
-                    Send
-                </Button>
-            </form>
-        </Modal>
+        <UserSendMessageModal bind:open={messaging} bind:person={person} />
     {/if}
-
-
 
 
     <StickyCard class="{$$props.class}">
@@ -164,7 +114,7 @@
                     <div>
                         <h1 class="flex flex-row">
                             <span class="font-bold text-lg">
-                                <UserLink badges user={person.person} showInstance={false} admin={is_admin}/>
+                                <UserLink badges user={person.person} showInstance={false} admin={is_admin} href/>
                             </span>
 
                             
@@ -221,9 +171,7 @@
                                     <!--- Actions for Logged-in <Users--->
                                     {#if $profile?.user && $profile.jwt && person.person.id != $profile.user.local_user_view.person.id}
                                         <!--- Message in Lemmy--->
-                                        <MenuButton
-                                            on:click={() => (messaging = true)}
-                                        >
+                                        <MenuButton on:click={() => (messaging = true)} >
                                             <Icon solid size="16" src={Envelope} />
                                             Message in Lemmy
                                         </MenuButton>