+++ src/lib/components/ui/modal/Modal.svelte
@@ -13,6 +13,7 @@
 
     export let fullHeight:boolean = false
     export let height:string = 'h-auto'
+    export let width:string = 'max-w-4xl'
     
     const dispatcher = createEventDispatcher()
 </script>
@@ -27,11 +28,11 @@
         <div transition:scale={{ start: 0.5, easing: expoOut }}
             class="overflow-y-auto rounded-xl max-w-full box-border w-full {fullHeight ? 'h-[95svh]' : height} {$$props.class}"
         >
-            <div class="w-full dark:!bg-zinc-950 rounded-xl max-w-4xl box-border mx-auto {fullHeight ? 'h-full' : height}">
+            <div class="w-full dark:!bg-zinc-950 rounded-xl {width} box-border mx-auto {fullHeight ? 'h-full' : height}">
                 <div
                     class="flex flex-col gap-4 p-3 rounded-xl overflow-none  w-full 
                     dark:bg-zinc-950 dark:border-zinc-800
-                    bg-white border border-slate-200  {fullHeight ? 'h-full' : 'h-auto'}"
+                    bg-white border border-slate-200  {fullHeight ? 'h-[95vh]' : 'h-auto'}"
                     class:rounded-b-none={action}
                     class:border-b-0={action}
                 >