+++ src/lib/lemmy/user.ts
@@ -51,7 +51,11 @@ export const addAdmin = async (handle: string, added: boolean, jwt: string) =>
     })
 
 
-export const blockUser = async function (personID: number, confirm:boolean=false):Promise<void> {
+export const blockUser = async function (personID: number, confirm:boolean=false, block?:boolean):Promise<boolean> {
+    const userProfile = get(profile)
+    if (!userProfile?.user || !userProfile?.jwt) throw new Error('Unauthenticated')
+    const blocked = isBlocked(userProfile.user, personID)
+    
     if (!confirm) {
         toast({
             title: "Confirmation",
@@ -59,21 +63,16 @@ export const blockUser = async function (personID: number, confirm:boolean=false
             type: "warning",
             action: async () => await blockUser(personID, true)
         })
-        return
+        return false
     }
     else {
         try {
-            const userProfile = get(profile)
-            if (!userProfile?.user || !userProfile?.jwt) throw new Error('Unauthenticated')
-
-            const blocked = isBlocked(userProfile.user, personID)
-
-            await getClient().blockPerson({
-                block: !blocked,
+            let blockResponse = await getClient().blockPerson({
+                block: block ?? !blocked,
                 person_id: personID,
             })
 
-            if (blocked) {
+            if (blockResponse?.blocked) {
                 const index = userProfile.user.person_blocks
                     .map((p) => p.target.id)
                     .indexOf(personID)
@@ -82,23 +81,19 @@ export const blockUser = async function (personID: number, confirm:boolean=false
             
             toast({
                 title: "Succcess",
-                content: `Successfully ${blocked ? 'unblocked' : 'blocked'} that user.`,
+                content: `Successfully ${blockResponse.blocked ? 'blocked' : 'unblocked'} that user.`,
                 type: 'success',
             })
 
-            return
-
-            // Refresh the page to effect the change in block status
-            //goto(window.location.href, {
-            //    invalidateAll: true,
-            //})
+            return blockResponse.blocked
 
         } catch (err) {
             toast({
                 content: err as any,
                 type: 'error',
+                title: 'Error'
             })
-            return
+            return blocked
         }
     }
 }